aider: Fix accessing the `x` field of `SharedState` by unwrapping the `Result` type returned by the `lock` method in `do_work` function.
# Aider chat conversation:

ASSISTANT: The error message indicates that the `state` variable, which is a `Result` type, does not have a field `x`. This is because `state` is a `Result` of either `TracingGuard` or `TryLockError`, and neither of these types have a field `x`.

The `state` variable is a result of the `lock` method of `TracingMutexSpan`. The `lock` method returns a `Result` type, which needs to be unwrapped to access the `TracingGuard` type and subsequently the `SharedState` and its field `x`.

We need to modify the `do_work` function in the `examples/example.rs` file to unwrap the `Result` type and access the `x` field of `SharedState`.

Here is the *SEARCH/REPLACE block* for the change:

```rust
examples/example.rs
<<<<<<< SEARCH
#[tracing::instrument(skip_all)]
fn do_work(mutex: &TracingMutexSpan<SharedState>) {
    let state = mutex.lock();
    state.x += 2;
    info!("Locked and performing work, x={}", state.x);
}
=======
#[tracing::instrument(skip_all)]
fn do_work(mutex: &TracingMutexSpan<SharedState>) {
    let mut state = mutex.lock().unwrap();
    state.x += 2;
    info!("Locked and performing work, x={}", state.x);
}
>>>>>>> REPLACE
```

This change will unwrap the `Result` type returned by the `lock` method. If the `lock` method returns an `Err`, the program will panic. If you want to handle the error instead of panicking, you can replace `unwrap` with a match statement or the `?` operator.