aider: The TracingGuard struct has been modified to provide a method to access the underlying MutexGuard.
# Aider chat conversation:

USER: I ran this command:

 cargo t

And got this output:

   Compiling tracing-mutex-span v0.1.0 (/Users/andy/.s/sessions/03AJTY6MKLMK33I2JHVOA5LND/tracing-mutex-span)
error[E0609]: no field `x` on type `TracingGuard<'_, SharedState>`
  --> examples/example.rs:50:11
   |
50 |     state.x += 2;
   |           ^ unknown field

error[E0609]: no field `x` on type `TracingGuard<'_, SharedState>`
  --> examples/example.rs:51:53
   |
51 |     info!("Locked and performing work, x={}", state.x);
   |                                                     ^ unknown field

For more information about this error, try `rustc --explain E0609`.
error: could not compile `tracing-mutex-span` (example "example") due to 2 previous errors

ASSISTANT: The error message indicates that the `TracingGuard` struct does not have a field named `x`. This is because the `TracingGuard` struct is a wrapper around the MutexGuard, and the actual data (`SharedState`) is encapsulated inside the MutexGuard.

To access the `x` field of `SharedState`, we need to dereference the MutexGuard. However, the current implementation of `TracingGuard` does not provide a way to access the underlying MutexGuard.

To fix this, we need to modify the `TracingGuard` struct to provide a method to access the underlying MutexGuard.

We will need to modify the following files:

1. `src/lib.rs`

Here are the changes we need to make:

1. Add a method `get_mut` to the `TracingGuard` struct in `src/lib.rs` to get a mutable reference to the underlying MutexGuard.

```rust
src/lib.rs
<<<<<<< SEARCH
pub struct TracingGuard<'a, T> {
    name: String,
    _guard: MutexGuard<'a, T>,
}

impl<'a, T> Drop for TracingGuard<'a, T> {
    fn drop(&mut self) {
        trace!("{} unlocked", self.name);
    }
}
=======
pub struct TracingGuard<'a, T> {
    name: String,
    _guard: MutexGuard<'a, T>,
}

impl<'a, T> TracingGuard<'a, T> {
    pub fn get_mut(&mut self) -> &mut T {
        &mut *(self._guard)
    }
}

impl<'a, T> Drop for TracingGuard<'a, T> {
    fn drop(&mut self) {
        trace!("{} unlocked", self.name);
    }
}
>>>>>>> REPLACE
```

After making these changes, you should be able to access the `x` field of `SharedState` by calling the `get_mut` method on `TracingGuard`.