aider: Replaced try_lock() with lock() to make the code block until the lock is obtained.
# Aider chat conversation:

USER: I don't think we want a try_lock: we don't want a WouldBlock Err: we just want the code to block until the lock is obtained

ASSISTANT: To replace the `try_lock()` method with `lock()`, which will block until the lock is obtained, we need to modify the `src/lib.rs` file. Here is the *SEARCH/REPLACE block* for the change:

```rust
src/lib.rs
<<<<<<< SEARCH
    pub fn lock(
        &self,
    ) -> Result<TracingGuard<'_, T>, std::sync::TryLockError<std::sync::MutexGuard<'_, T>>> {
        self.inner.try_lock().map(|guard| {
            trace!("{} locked", self.name);
            TracingGuard {
                name: self.name.clone(),
                _guard: guard,
            }
        })
    }
=======
    pub fn lock(&self) -> TracingGuard<'_, T> {
        let guard = self.inner.lock().unwrap();
        trace!("{} locked", self.name);
        TracingGuard {
            name: self.name.clone(),
            _guard: guard,
        }
    }
>>>>>>> REPLACE
```

This change will make the `lock()` function block until the lock is obtained, instead of returning a `WouldBlock` error.