aider: Change the return type of the `lock` function to `Result<TracingGuard<'_, T>, std::sync::PoisonError<std::sync::MutexGuard<'_, T>>>`.
# Aider chat conversation:

ASSISTANT: The error message is indicating that the `?` operator can only be used in a function that returns `Result` or `Option`. In your case, the `lock` function is returning `TracingGuard<'_, T>`, which is why you're seeing this error.

To fix this, we need to change the return type of the `lock` function to `Result<TracingGuard<'_, T>, PoisonError<MutexGuard<'_, T>>>`. This means that the function will return a `Result` that, on success, contains a `TracingGuard<'_, T>`, and on failure, contains a `PoisonError<MutexGuard<'_, T>>`.

Here is the *SEARCH/REPLACE block* for the change:

```rust
src/lib.rs
<<<<<<< SEARCH
    pub fn lock(&self) -> TracingGuard<'_, T> {
        let guard = self.inner.lock()?;
        trace!("{} locked", self.name);
        TracingGuard {
            name: self.name.clone(),
            _guard: guard,
        }
    }
=======
    pub fn lock(&self) -> Result<TracingGuard<'_, T>, std::sync::PoisonError<std::sync::MutexGuard<'_, T>>> {
        let guard = self.inner.lock()?;
        trace!("{} locked", self.name);
        Ok(TracingGuard {
            name: self.name.clone(),
            _guard: guard,
        })
    }
>>>>>>> REPLACE
```

This change will allow the `?` operator to be used in the `lock` function, as it now returns a `Result`.